python_extractor/requirements.txt
torch
torchvision
numpy
Pillow
tqdm